+++ src/main/java/lagatrix/server/manager/GPUManager.java
@@ -1,8 +1,8 @@
 package lagatrix.server.manager;
 
 import lagatrix.server.entities.devices.GPU;
-import lagatrix.server.exceptions.manager.GPUException;
-import lagatrix.server.manager.information.GPUInfo;
+import lagatrix.server.exceptions.manager.hardware.GPUException;
+import lagatrix.server.manager.information.hardware.GPUInfo;
 import lagatrix.server.tools.command.CommandExecutor;
 
 /**