+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - LASRV003 - 23/03/2023 - Add the RAMManager class to obtain the RAM.
 - LASRV004 - 23/03/2023 - Add the PartitonInfo class to obtain information of partitions.
 - LASRV004 - 24/03/2023 - Add the PartitionManager class to obtain partitions.
+- LASRV005 - 24/03/2024 - Add the UnameInfo class to obtain hostame and kernel.
 
 ### Fixed
 - LASRV001 - 21/03/2023 - Refactor code in the temperature detectors.